move blockchain creation code to execution
