+++ arbnode/node.go
@@ -22,10 +22,8 @@ import (
 	"github.com/ethereum/go-ethereum/core"
 	"github.com/ethereum/go-ethereum/core/rawdb"
 	"github.com/ethereum/go-ethereum/core/types"
-	"github.com/ethereum/go-ethereum/core/vm"
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/eth"
-	"github.com/ethereum/go-ethereum/eth/ethconfig"
 	"github.com/ethereum/go-ethereum/eth/filters"
 	"github.com/ethereum/go-ethereum/ethdb"
 	"github.com/ethereum/go-ethereum/log"
@@ -34,9 +32,6 @@ import (
 	"github.com/ethereum/go-ethereum/rpc"
 
 	"github.com/offchainlabs/nitro/arbnode/execution"
-	"github.com/offchainlabs/nitro/arbos"
-	"github.com/offchainlabs/nitro/arbos/arbosState"
-	"github.com/offchainlabs/nitro/arbstate"
 	"github.com/offchainlabs/nitro/arbutil"
 	"github.com/offchainlabs/nitro/broadcastclient"
 	"github.com/offchainlabs/nitro/broadcastclients"
@@ -47,7 +42,6 @@ import (
 	"github.com/offchainlabs/nitro/solgen/go/ospgen"
 	"github.com/offchainlabs/nitro/solgen/go/rollupgen"
 	"github.com/offchainlabs/nitro/staker"
-	"github.com/offchainlabs/nitro/statetransfer"
 	"github.com/offchainlabs/nitro/util/contracts"
 	"github.com/offchainlabs/nitro/util/headerreader"
 	"github.com/offchainlabs/nitro/util/signature"
@@ -416,7 +410,7 @@ type Config struct {
 	Wasm                   WasmConfig                  `koanf:"wasm"`
 	SyncMonitor            SyncMonitorConfig           `koanf:"sync-monitor"`
 	Dangerous              DangerousConfig             `koanf:"dangerous"`
-	Caching                CachingConfig               `koanf:"caching"`
+	Caching                execution.CachingConfig     `koanf:"caching"`
 	Archive                bool                        `koanf:"archive"`
 	TxLookupLimit          uint64                      `koanf:"tx-lookup-limit"`
 	TransactionStreamer    TransactionStreamerConfig   `koanf:"transaction-streamer" reload:"hot"`
@@ -482,7 +476,7 @@ func ConfigAddOptions(prefix string, f *flag.FlagSet, feedInputEnable bool, feed
 	WasmConfigAddOptions(prefix+".wasm", f)
 	SyncMonitorConfigAddOptions(prefix+".sync-monitor", f)
 	DangerousConfigAddOptions(prefix+".dangerous", f)
-	CachingConfigAddOptions(prefix+".caching", f)
+	execution.CachingConfigAddOptions(prefix+".caching", f)
 	f.Uint64(prefix+".tx-lookup-limit", ConfigDefault.TxLookupLimit, "retain the ability to lookup transactions by hash for the past N blocks (0 = all blocks)")
 	TransactionStreamerConfigAddOptions(prefix+".transaction-streamer", f)
 
@@ -509,7 +503,7 @@ var ConfigDefault = Config{
 	Dangerous:              DefaultDangerousConfig,
 	Archive:                false,
 	TxLookupLimit:          126_230_400, // 1 year at 4 blocks per second
-	Caching:                DefaultCachingConfig,
+	Caching:                execution.DefaultCachingConfig,
 	TransactionStreamer:    DefaultTransactionStreamerConfig,
 }
 
@@ -620,33 +614,6 @@ func (w *WasmConfig) FindMachineDir() (string, bool) {
 		}
 	}
 	return "", false
-}
-
-type CachingConfig struct {
-	Archive               bool          `koanf:"archive"`
-	BlockCount            uint64        `koanf:"block-count"`
-	BlockAge              time.Duration `koanf:"block-age"`
-	TrieTimeLimit         time.Duration `koanf:"trie-time-limit"`
-	TrieDirtyCache        int           `koanf:"trie-dirty-cache"`
-	SnapshotRestoreMaxGas uint64        `koanf:"snapshot-restore-gas-limit"`
-}
-
-func CachingConfigAddOptions(prefix string, f *flag.FlagSet) {
-	f.Bool(prefix+".archive", DefaultCachingConfig.Archive, "retain past block state")
-	f.Uint64(prefix+".block-count", DefaultCachingConfig.BlockCount, "minimum number of recent blocks to keep in memory")
-	f.Duration(prefix+".block-age", DefaultCachingConfig.BlockAge, "minimum age a block must be to be pruned")
-	f.Duration(prefix+".trie-time-limit", DefaultCachingConfig.TrieTimeLimit, "maximum block processing time before trie is written to hard-disk")
-	f.Int(prefix+".trie-dirty-cache", DefaultCachingConfig.TrieDirtyCache, "amount of memory in megabytes to cache state diffs against disk with (larger cache lowers database growth)")
-	f.Uint64(prefix+".snapshot-restore-gas-limit", DefaultCachingConfig.SnapshotRestoreMaxGas, "maximum gas rolled back to recover snapshot")
-}
-
-var DefaultCachingConfig = CachingConfig{
-	Archive:               false,
-	BlockCount:            128,
-	BlockAge:              30 * time.Minute,
-	TrieTimeLimit:         time.Hour,
-	TrieDirtyCache:        1024,
-	SnapshotRestoreMaxGas: 300_000_000_000,
 }
 
 type Node struct {
@@ -1365,133 +1332,4 @@ func CreateDefaultStackForTest(dataDir string) (*node.Node, error) {
 		return nil, fmt.Errorf("error creating protocol stack: %w", err)
 	}
 	return stack, nil
-}
-
-func DefaultCacheConfigFor(stack *node.Node, cachingConfig *CachingConfig) *core.CacheConfig {
-	baseConf := ethconfig.Defaults
-	if cachingConfig.Archive {
-		baseConf = ethconfig.ArchiveDefaults
-	}
-
-	return &core.CacheConfig{
-		TrieCleanLimit:        baseConf.TrieCleanCache,
-		TrieCleanJournal:      stack.ResolvePath(baseConf.TrieCleanCacheJournal),
-		TrieCleanRejournal:    baseConf.TrieCleanCacheRejournal,
-		TrieCleanNoPrefetch:   baseConf.NoPrefetch,
-		TrieDirtyLimit:        cachingConfig.TrieDirtyCache,
-		TrieDirtyDisabled:     cachingConfig.Archive,
-		TrieTimeLimit:         cachingConfig.TrieTimeLimit,
-		TriesInMemory:         cachingConfig.BlockCount,
-		TrieRetention:         cachingConfig.BlockAge,
-		SnapshotLimit:         baseConf.SnapshotCache,
-		Preimages:             baseConf.Preimages,
-		SnapshotRestoreMaxGas: cachingConfig.SnapshotRestoreMaxGas,
-	}
-}
-
-func WriteOrTestGenblock(chainDb ethdb.Database, initData statetransfer.InitDataReader, chainConfig *params.ChainConfig, accountsPerSync uint) error {
-	arbstate.RequireHookedGeth()
-
-	EmptyHash := common.Hash{}
-	prevHash := EmptyHash
-	prevDifficulty := big.NewInt(0)
-	blockNumber, err := initData.GetNextBlockNumber()
-	if err != nil {
-		return err
-	}
-	storedGenHash := rawdb.ReadCanonicalHash(chainDb, blockNumber)
-	timestamp := uint64(0)
-	if blockNumber > 0 {
-		prevHash = rawdb.ReadCanonicalHash(chainDb, blockNumber-1)
-		if prevHash == EmptyHash {
-			return fmt.Errorf("block number %d not found in database", chainDb)
-		}
-		prevHeader := rawdb.ReadHeader(chainDb, prevHash, blockNumber-1)
-		if prevHeader == nil {
-			return fmt.Errorf("block header for block %d not found in database", chainDb)
-		}
-		timestamp = prevHeader.Time
-	}
-	stateRoot, err := arbosState.InitializeArbosInDatabase(chainDb, initData, chainConfig, timestamp, accountsPerSync)
-	if err != nil {
-		return err
-	}
-
-	genBlock := arbosState.MakeGenesisBlock(prevHash, blockNumber, timestamp, stateRoot, chainConfig)
-	blockHash := genBlock.Hash()
-
-	if storedGenHash == EmptyHash {
-		// chainDb did not have genesis block. Initialize it.
-		core.WriteHeadBlock(chainDb, genBlock, prevDifficulty)
-		log.Info("wrote genesis block", "number", blockNumber, "hash", blockHash)
-	} else if storedGenHash != blockHash {
-		return fmt.Errorf("database contains data inconsistent with initialization: database has genesis hash %v but we built genesis hash %v", storedGenHash, blockHash)
-	} else {
-		log.Info("recreated existing genesis block", "number", blockNumber, "hash", blockHash)
-	}
-
-	return nil
-}
-
-func TryReadStoredChainConfig(chainDb ethdb.Database) *params.ChainConfig {
-	EmptyHash := common.Hash{}
-
-	block0Hash := rawdb.ReadCanonicalHash(chainDb, 0)
-	if block0Hash == EmptyHash {
-		return nil
-	}
-	return rawdb.ReadChainConfig(chainDb, block0Hash)
-}
-
-func WriteOrTestChainConfig(chainDb ethdb.Database, config *params.ChainConfig) error {
-	EmptyHash := common.Hash{}
-
-	block0Hash := rawdb.ReadCanonicalHash(chainDb, 0)
-	if block0Hash == EmptyHash {
-		return errors.New("block 0 not found")
-	}
-	storedConfig := rawdb.ReadChainConfig(chainDb, block0Hash)
-	if storedConfig == nil {
-		rawdb.WriteChainConfig(chainDb, block0Hash, config)
-		return nil
-	}
-	height := rawdb.ReadHeaderNumber(chainDb, rawdb.ReadHeadHeaderHash(chainDb))
-	if height == nil {
-		return errors.New("non empty chain config but empty chain")
-	}
-	err := storedConfig.CheckCompatible(config, *height)
-	if err != nil {
-		return err
-	}
-	rawdb.WriteChainConfig(chainDb, block0Hash, config)
-	return nil
-}
-
-func GetBlockChain(chainDb ethdb.Database, cacheConfig *core.CacheConfig, chainConfig *params.ChainConfig, nodeConfig *Config) (*core.BlockChain, error) {
-	engine := arbos.Engine{
-		IsSequencer: true,
-	}
-
-	vmConfig := vm.Config{
-		EnablePreimageRecording: false,
-	}
-
-	return core.NewBlockChain(chainDb, cacheConfig, chainConfig, engine, vmConfig, shouldPreserveFalse, &nodeConfig.TxLookupLimit)
-}
-
-func WriteOrTestBlockChain(chainDb ethdb.Database, cacheConfig *core.CacheConfig, initData statetransfer.InitDataReader, chainConfig *params.ChainConfig, nodeConfig *Config, accountsPerSync uint) (*core.BlockChain, error) {
-	err := WriteOrTestGenblock(chainDb, initData, chainConfig, accountsPerSync)
-	if err != nil {
-		return nil, err
-	}
-	err = WriteOrTestChainConfig(chainDb, chainConfig)
-	if err != nil {
-		return nil, err
-	}
-	return GetBlockChain(chainDb, cacheConfig, chainConfig, nodeConfig)
-}
-
-// Don't preserve reorg'd out blocks
-func shouldPreserveFalse(_ *types.Header) bool {
-	return false
 }
